add flags
