+++ src/InitOMPasses.hpp
@@ -11,6 +11,12 @@ namespace onnx_mlir {
 void initOMPasses() {
   // All passes implemented within onnx-mlir should register within this
   // function to make themselves available as a command-line option.
+  mlir::registerPass("onnx-graph-optimize",
+      "Invoke other passes iteratively to optimize ONNX graph.",
+      []() -> std::unique_ptr<mlir::Pass> {
+        return mlir::createONNXGraphOptimizePass();
+      });
+
   mlir::registerPass("decompose-onnx",
       "Decompose ONNX operations into composition of other ONNX operations.",
       []() -> std::unique_ptr<mlir::Pass> {
