+++ src/Pass/Passes.hpp
@@ -21,6 +21,7 @@ class Pass;
 
 /// Pass for ONNX graph level optimization
 std::unique_ptr<Pass> createONNXGraphOptimizePass();
+std::unique_ptr<Pass> createONNXGraphOptimizePass(int threshold);
 
 /// Pass for rewriting inside frontend dialect.
 std::unique_ptr<Pass> createDecomposeONNXToONNXPass();
