+++ src/Support/OMOptions.cpp
@@ -45,3 +45,7 @@ llvm::cl::opt<bool> disableMemoryBundling("disable-memory-bundling",
                    "Buffer management by MLIR is used instead.\n"
                    "Try this if you experience a significant compile time."),
     llvm::cl::init(false), llvm::cl::cat(OMPassOptions));
+
+llvm::cl::opt<int> onnxGraphOptimizeThreshold("onnx-graph-optimize-threshold",
+    llvm::cl::desc(" max iteration for graph optimization passes."),
+    llvm::cl::init(3), llvm::cl::cat(OMPassOptions));
