+++ src/Support/OMOptions.hpp
@@ -24,3 +24,5 @@ extern llvm::cl::OptionCategory OMPassOptions;
 extern llvm::cl::opt<std::string> instrumentONNXOps;
 
 extern llvm::cl::opt<bool> disableMemoryBundling;
+
+extern llvm::cl::opt<int> onnxGraphOptimizeThreshold;
